+++ RELEASES.md
@@ -48,6 +48,8 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: generate linker tables
 * prov-template: fully qualified names for outputs
 * provconvert: support for template library path
+* ProvToolbox: upgrade to jackson 2.18.5, and then 2.19.4
+
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java